chore(web): Extract google login functionality into dedicated components
