+++ packages/web/src/views/Login/Login.tsx
@@ -1,14 +1,13 @@
 import React, { useEffect, useRef, useState } from "react";
-import GoogleButton from "react-google-button";
 import { useNavigate } from "react-router-dom";
-import { v4 as uuidv4 } from "uuid";
-import { useGoogleLogin } from "@react-oauth/google";
 import { useAuthCheck } from "@web/auth/useAuthCheck";
 import { AuthApi } from "@web/common/apis/auth.api";
 import { WaitlistApi } from "@web/common/apis/waitlist.api";
 import { ROOT_ROUTES } from "@web/common/constants/routes";
 import { AlignItems, FlexDirections } from "@web/components/Flex/styled";
 import { LoginAbsoluteOverflowLoader } from "@web/components/LoginAbsoluteOverflowLoader/LoginAbsoluteOverflowLoader";
+import { GoogleButton } from "@web/components/oauth/google/GoogleButton";
+import { useGoogleLogin } from "@web/components/oauth/google/useGoogleLogin";
 import {
   ActionButton,
   Card,
@@ -48,7 +47,6 @@ export const LoginView = () => {
   const [apiError, setApiError] = useState<string | null>(null);
 
   const { isAuthenticated: isAlreadyAuthenticated } = useAuthCheck();
-  const antiCsrfToken = useRef(uuidv4()).current;
 
   useEffect(() => {
     if (window.location.hostname === "localhost") {
@@ -73,50 +71,17 @@ export const LoginView = () => {
     }
   }, [isAuthenticated, navigate]);
 
-  const SCOPES_REQUIRED = [
-    "email",
-    "https://www.googleapis.com/auth/calendar.readonly",
-    "https://www.googleapis.com/auth/calendar.events",
-  ];
-
-  const isMissingPermissions = (scope: string) => {
-    const scopesGranted = scope.split(" ");
-    return SCOPES_REQUIRED.some((s) => !scopesGranted.includes(s));
-  };
-
   const startLoginFlow = useGoogleLogin({
-    flow: "auth-code",
-    scope: SCOPES_REQUIRED.join(" "),
-    state: antiCsrfToken,
-    onSuccess: async ({ code, scope, state }) => {
-      const isFromHacker = state !== antiCsrfToken;
-      if (isFromHacker) {
-        alert("Nice try, hacker");
-        return;
-      }
-
-      if (isMissingPermissions(scope)) {
-        alert("Missing permissions, please click all the checkboxes");
-        return;
-      }
-
-      setIsAuthenticating(true);
-      try {
-        await AuthApi.loginOrSignup(code);
-        setIsAuthenticated(true);
-      } catch (e) {
-        console.error(e);
-        alert("Login failed. Please try again.");
-      } finally {
-        setIsAuthenticating(false);
-      }
+    onStart: () => setIsAuthenticating(true),
+    onSuccess: async (code) => {
+      await AuthApi.loginOrSignup(code);
+      setIsAuthenticated(true);
     },
     onError: (error) => {
-      alert(`Login failed because: ${error.error}`);
       console.error(error);
+      setIsAuthenticating(false);
     },
   });
-
   const handleCheckWaitlistSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
     const processedInput = emailInput.trim().toLowerCase();
@@ -281,8 +246,6 @@ export const LoginView = () => {
                     </InfoText>
                     <SignInButtonWrapper>
                       <GoogleButton
-                        aria-label="Sign in with Google"
-                        type="light"
                         onClick={handleButtonClick}
                         disabled={isAuthenticating}
                       />
